+++ lib/packages/contracts/src/generateContract.ts
@@ -1,4 +1,13 @@
 /** @module contracts */
+
+/**
+ * Copyright (c) 2019 Burst Apps Team
+ *
+ * Credits to AJ ONeal for the two-complements stuff
+ * https://coolaj86.com/articles/convert-decimal-to-hex-with-js-bigints/
+ */
+
+
 import {GenerateContractArgs} from './typings/args';
 import {generateContractBytes} from './internal/generateContractBytes';
 import {convertByteArrayToHexString} from '@burstjs/util';